requirements.txt
Flask==2.2.3
pyserial==3.5
Werkzeug==2.2.3
MarkupSafe==2.1.2
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3